com: allow RTS and DTR to be set; only open port once
Allow RTS# and DTR# status to be set from the command line.
Only open the port once; there is no need to re-open it to
change the serial port settings. Instead, factor out the
setting of COM port parameters from COM port open.

To avoid tacking on parameters to function calls indefinitely in the
future, stash the com port parameters into a structure. This could
also be augmented to hold the local state of the port (e.g. fd) rather
than having that be a static variable in the future.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>